\name{03_constants}
\alias{ZERO-class}
\alias{ZERO}
\alias{sqps}
\alias{cubps}
\title{Constant Values}
\description{The ZERO constant for \link{NestMatrix} and the sqps and cubps constants, for the \link{polyplot} and \link{cubplot} functions.}
\usage{
ZERO

sqps
cubps
}
\seealso{
\link{NestMatrix}

\link{polyplot}, \link{cubplot}
}
\examples{
sqps
cubps
}
